moved category service to its own folder
