+++ client-side/src/app/portals/Caliber/caliber.module.ts
@@ -161,7 +161,7 @@ import { TagFilterPipe } from './settings/screening/question/question.filter';
 import {BucketFilterPipe} from './settings/screening/skillType-buckets/skillType-buckets.filter';
 import { PDFService } from './services/pdf.service';
 import { ReportingService } from './services/reporting.service';
-import { CategoryService } from './services/category.service';
+import { CategoryService } from '../Caliber/services/category/category.service';
 
 
 @NgModule({
