+++ client-side/src/app/portals/Caliber/caliber.test.module.ts
@@ -166,7 +166,7 @@ import { TagFilterPipe } from './settings/screening/question/question.filter';
 import {BucketFilterPipe} from './settings/screening/skillType-buckets/skillType-buckets.filter';
 import { PDFService } from './services/pdf.service';
 import { ReportingService } from './services/reporting.service';
-import { CategoryService } from './services/category.service';
+import { CategoryService } from './services/category/category.service';
 
 
 export const Dependencies = {
