--- client-side/src/app/portals/Caliber/services/category.service.spec.ts
@@ -1,23 +1,0 @@
-import { TestBed, inject } from '@angular/core/testing';
-import { HttpClientModule, HttpClient} from '@angular/common/http';
-import { AlertsService } from '../services/alerts.service';
-import { CategoryService } from './category.service';
-
-xdescribe('CategoryService', () => {
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      imports: [
-        HttpClientModule
-      ],
-      providers: [
-        CategoryService,
-        HttpClient,
-        AlertsService
-      ]
-    });
-  });
-
-  it('should be created', inject([CategoryService], (service: CategoryService) => {
-    expect(service).toBeTruthy();
-  }));
-});
